some compatibility updates for numpy 2.0

+++ src/qrisp/simulator/unitary_management.py
@@ -31,7 +31,7 @@ np_dtype = np.complex64
 id_matrix = np.eye(2, dtype=np_dtype)
 
 pauli_x = np.asarray([[0, 1], [1, 0]], dtype=np_dtype)
-pauli_y = (np.asarray([[0, 0], [0, 0]]) + 1j * np.asarray([[0, -1], [1, 0]])).astype(np_dtype)
+pauli_y = (np.asarray([[0, 0], [0, 0]], dtype = np_dtype) + 1j * np.asarray([[0, -1], [1, 0]])).astype(np_dtype)
 pauli_z = np.asarray([[1, 0], [0, -1]], dtype=np_dtype)
 from sympy.core.expr import Expr
 import numpy
@@ -63,7 +63,7 @@ def u3matrix(theta, phi, lam, global_phase, use_sympy = False):
     res[1, 0] = module.exp(I * phi) * module.sin(theta / 2)
     res[1, 1] = module.exp(I * (phi + lam)) * module.cos(theta / 2)
 
-    return res * module.exp(I * global_phase)
+    return res * module.exp(I * global_phase, dtype = np_dtype)
 
 
 # Efficient function to generate the unitary of a controlled gate
